chore: move utils to utils folder
