+++ packages/frontend/src/components/common/Filters/FilterInputs/DateFilterInputs.tsx
@@ -13,8 +13,8 @@ import {
 import { Flex, NumberInput, Text } from '@mantine/core';
 import React from 'react';
 import { useFiltersContext } from '../FiltersProvider';
+import { getFirstDayOfWeek } from '../utils/dateUtils';
 import { getPlaceholderByFilterTypeAndOperator } from '../utils/getPlaceholderByFilterTypeAndOperator';
-import { getFirstDayOfWeek } from './dateUtils';
 import DefaultFilterInputs, { FilterInputsProps } from './DefaultFilterInputs';
 import FilterDatePicker from './FilterDatePicker';
 import FilterDateRangePicker from './FilterDateRangePicker';
